[data] imoroved fields types in CoinApiResponse

+++ data/src/main/java/com/davidepani/data/models/CoinApiResponse.kt
@@ -3,7 +3,7 @@ package com.davidepani.data.models
 import com.google.gson.annotations.SerializedName
 
 data class CoinApiResponse(
-    val ath: Int,
+    val ath: Double,
     @SerializedName("ath_change_percentage")
     val athChangePercentage: Double,
     @SerializedName("ath_date")
@@ -14,19 +14,19 @@ data class CoinApiResponse(
     @SerializedName("atl_date")
     val atlDate: String,
     @SerializedName("circulating_supply")
-    val circulatingSupply: Int,
+    val circulatingSupply: Double,
     @SerializedName("current_price")
-    val currentPrice: Int,
+    val currentPrice: Double,
     @SerializedName("fully_diluted_valuation")
     val fullyDilutedValuation: Long,
     @SerializedName("high_24h")
-    val high24h: Int,
+    val high24h: Double,
     val id: String,
     val image: String,
     @SerializedName("last_updated")
     val lastUpdated: String,
     @SerializedName("low_24h")
-    val low24h: Int,
+    val low24h: Double,
     @SerializedName("market_cap")
     val marketCap: Long,
     @SerializedName("market_cap_change_24h")
@@ -36,7 +36,7 @@ data class CoinApiResponse(
     @SerializedName("market_cap_rank")
     val marketCapRank: Int,
     @SerializedName("max_supply")
-    val maxSupply: Int,
+    val maxSupply: Double,
     val name: String,
     @SerializedName("price_change_24h")
     val priceChange24h: Double,
@@ -45,7 +45,7 @@ data class CoinApiResponse(
     val roi: Any,
     val symbol: String,
     @SerializedName("total_supply")
-    val totalSupply: Int,
+    val totalSupply: Double,
     @SerializedName("total_volume")
-    val totalVolume: Long
+    val totalVolume: Int
 )
